Сделал рефакторинг трех разделов urls, forms и views в виде разделения на products и categories
