+++ server/products/forms/__init__.py
@@ -1,0 +1,2 @@
+from .categories import *
+from .products import *
