+++ server/products/forms/categories.py
@@ -1,6 +1,6 @@
 from django import forms
 
-from .models import Category, Product
+from products.models import Category
 
 
 class CategoryForm(forms.Form):
@@ -40,21 +40,3 @@ class CategoryModelForm(forms.ModelForm):
             )
         }
 
-
-class ProductModelForm(forms.ModelForm):
-    class Meta:
-        model = Product
-        fields = ['name', 'image', 'category', 'description', 'cost']
-        widgets = {
-            'name': forms.widgets.TextInput(
-                attrs={
-                    'class': 'model-form'
-                }
-            ),
-
-            'description': forms.widgets.Textarea(
-                attrs={
-                    'class': 'model-form'
-                }
-            )
-        }
