+++ server/products/templates/categories/create_cat.html
@@ -2,6 +2,10 @@
 
 {% load static %}
 
+{% block page_title %}
+Create category
+{% endblock %}
+
 {% block page_static %}
 {% endblock %}
