+++ server/products/templates/categories/update_cat.html
@@ -2,6 +2,10 @@
 
 {% load static %}
 
+{% block page_title %}
+Update category
+{% endblock %}
+
 {% block page_static %}
 {% endblock %}
