+++ server/products/templates/products/create_prod.html
@@ -2,6 +2,10 @@
 
 {% load static %}
 
+{% block page_title %}
+Create product
+{% endblock %}
+
 {% block page_static %}
 {% endblock %}
