+++ server/products/templates/products/update_prod.html
@@ -2,6 +2,10 @@
 
 {% load static %}
 
+{% block page_title %}
+Update product
+{% endblock %}
+
 {% block page_static %}
 {% endblock %}
