+++ server/server/templates/server/menu.html
@@ -5,28 +5,28 @@
             <span class="navigation-list__item-text">
                 Главная
             </span>
-            <a href="{% url 'main' %}" class="navigation-list__item-link">
+            <a href="{% url 'main:main' %}" class="navigation-list__item-link">
             </a>
         </li>
         <li class="navigation-list__item">
             <span class="navigation-list__item-text">
                 Каталог
             </span>
-            <a href="{% url 'list' %}" class="navigation-list__item-link">
+            <a href="{% url 'products:list' %}" class="navigation-list__item-link">
             </a>
         </li>
         <li class="navigation-list__item">
             <span class="navigation-list__item-text">
                 О проекте
             </span>
-            <a href="{% url 'about' %}" class="navigation-list__item-link">
+            <a href="{% url 'main:about' %}" class="navigation-list__item-link">
             </a>
         </li>
         <li class="navigation-list__item">
             <span class="navigation-list__item-text">
                 Контакты
             </span>
-            <a href="{% url 'contacts' %}" class="navigation-list__item-link">
+            <a href="{% url 'main:contacts' %}" class="navigation-list__item-link">
             </a>
         </li>
     </ul>
